# docs/contracts/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "CandlePointResponse": {
        "description": "Single candle row.",
        "properties": {
          "close": {
            "title": "Close",
            "type": "number"
          },
          "high": {
            "title": "High",
            "type": "number"
          },
          "low": {
            "title": "Low",
            "type": "number"
          },
          "open": {
            "title": "Open",
            "type": "number"
          },
          "time": {
            "title": "Time",
            "type": "string"
          },
          "volume": {
            "title": "Volume",
            "type": "number"
          }
        },
        "required": [
          "time",
          "open",
          "high",
          "low",
          "close",
          "volume"
        ],
        "title": "CandlePointResponse",
        "type": "object"
      },
      "CandlesResponse": {
        "additionalProperties": true,
        "description": "Candles endpoint response.",
        "properties": {
          "candles": {
            "items": {
              "$ref": "#/components/schemas/CandlePointResponse"
            },
            "title": "Candles",
            "type": "array"
          },
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "market_type": {
            "title": "Market Type",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "timeframe": {
            "title": "Timeframe",
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "market_type",
          "timeframe",
          "source",
          "count",
          "candles"
        ],
        "title": "CandlesResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ResearchQuery": {
        "additionalProperties": false,
        "properties": {
          "operation": {
            "maxLength": 80,
            "minLength": 1,
            "pattern": "^[a-z0-9_.]+$",
            "title": "Operation",
            "type": "string"
          },
          "params": {
            "additionalProperties": true,
            "maxProperties": 20,
            "title": "Params",
            "type": "object"
          }
        },
        "required": [
          "operation"
        ],
        "title": "ResearchQuery",
        "type": "object"
      },
      "SavedResearchWrite": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "operation": {
            "maxLength": 80,
            "minLength": 1,
            "pattern": "^[a-z0-9_.]+$",
            "title": "Operation",
            "type": "string"
          },
          "params": {
            "additionalProperties": true,
            "maxProperties": 20,
            "title": "Params",
            "type": "object"
          }
        },
        "required": [
          "operation",
          "name"
        ],
        "title": "SavedResearchWrite",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "7/24 Finansal Sinyal ve Analiz REST API",
    "title": "Otonom Analiz API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/borsapy/candles/{symbol}": {
      "get": {
        "operationId": "private_candles_borsapy_candles__symbol__get",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "maxLength": 25,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9]{1,20}(?:\\.IS)?$",
              "title": "Symbol",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "interval",
            "required": false,
            "schema": {
              "default": "1d",
              "enum": [
                "1m",
                "5m",
                "15m",
                "30m",
                "1h",
                "4h",
                "1d",
                "1wk",
                "1mo",
                "2d",
                "3d",
                "4d",
                "5d",
                "6d",
                "2wk",
                "3wk",
                "2mo",
                "3mo"
              ],
              "title": "Interval",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 1000,
              "maximum": 2000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Private Candles Borsapy Candles  Symbol  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Private Candles",
        "tags": [
          "Borsapy research"
        ],
        "x-access-role": "admin"
      }
    },
    "/borsapy/catalog": {
      "get": {
        "operationId": "research_catalog_borsapy_catalog_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Research Catalog Borsapy Catalog Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Research Catalog",
        "tags": [
          "Borsapy research"
        ],
        "x-access-role": "admin"
      }
    },
...
    "/borsapy/query": {
      "post": {
        "operationId": "research_query_borsapy_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResearchQuery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Research Query Borsapy Query Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Research Query",
        "tags": [
          "Borsapy research"
        ],
        "x-access-role": "admin"
      }
    },
    "/borsapy/saved": {
      "get": {
        "operationId": "list_saved_borsapy_saved_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response List Saved Borsapy Saved Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Saved",
        "tags": [
          "Borsapy research"
        ],
        "x-access-role": "admin"
      },
      "post": {
        "operationId": "create_saved_borsapy_saved_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavedResearchWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Saved Borsapy Saved Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Saved",
        "tags": [
          "Borsapy research"
        ],
        "x-access-role": "admin"
      }
    },
    "/borsapy/saved/{workspace_id}": {
      "delete": {
        "operationId": "delete_saved_borsapy_saved__workspace_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Delete Saved Borsapy Saved  Workspace Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Saved",
        "tags": [
          "Borsapy research"
        ],
        "x-access-role": "admin"
      },
      "put": {
        "operationId": "update_saved_borsapy_saved__workspace_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SavedResearchWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Update Saved Borsapy Saved  Workspace Id  Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Saved",
        "tags": [
          "Borsapy research"
        ],
        "x-access-role": "admin"
      }
    },
...
    "/candles/{symbol}": {
      "get": {
        "operationId": "get_candles_candles__symbol__get",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "default": "borsapy",
              "pattern": "^(borsapy|legacy)$",
              "title": "Source",
              "type": "string"
            }
          },
          {
            "description": "Piyasa türü (BIST/Kripto)",
            "in": "query",
            "name": "market_type",
            "required": false,
            "schema": {
              "default": "BIST",
              "description": "Piyasa türü (BIST/Kripto)",
              "title": "Market Type",
              "type": "string"
            }
          },
          {
            "description": "Timeframe (15m, 30m, 1h, 2h, 4h, 8h, 12h, 1d, 2d, 3d, 4d, 5d, 6d, 1wk, 2wk, 3wk, 1mo, 2mo, 3mo)",
            "in": "query",
            "name": "timeframe",
            "required": false,
            "schema": {
              "default": "1d",
              "description": "Timeframe (15m, 30m, 1h, 2h, 4h, 8h, 12h, 1d, 2d, 3d, 4d, 5d, 6d, 1wk, 2wk, 3wk, 1mo, 2mo, 3mo)",
              "title": "Timeframe",
              "type": "string"
            }
          },
          {
            "description": "Number of candles (max 2000)",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 500,
              "description": "Number of candles (max 2000)",
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandlesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Candles",
        "tags": [
          "Market Data"
        ],
        "x-access-role": "admin"
      }
    },
...
  }
}